CMake: Export targets for use in other projects
This commit adds the modern CMake approach to handling inter-project
dependencies: A configuration file describing the exported libraries
and headers will be generated which can be imported in other
projects using find_package.

Note that the collision checker is excluded from this mechanism for now
since the way its include directories are handled most likely wouldn't
work with this approach.